Fix tests warnings Improve checklist detailed html page Before applying a a command check if its possible to execute it(Eg. when creating a tag for a checklist, first we need to validate if the checklist exist) Throw exception when the command cant execute

+++ src/main/java/pt/isel/ls/Commands/DeleteTagsGid.java
@@ -2,23 +2,35 @@ package pt.isel.ls.Commands;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.util.HashMap;
 
 public class DeleteTagsGid extends Command {
     private final String method = "DELETE";
     private final String[] path = {"", "tags", "{gid}"};
+
     @Override
     public Object execute(HashMap<String, String> params, Connection con) throws Exception {
-        String s1 = "delete from tag_checklist where Tg_id = ?";
+        String s1 = "select * from tag where Tg_id = ?";
+
         PreparedStatement ps = con.prepareStatement(s1);
         ps.setString(1, params.get("{gid}"));
-        ps.executeUpdate();
+        ResultSet rs = ps.executeQuery();
 
-        s1 = "delete from tag where Tg_id = ?";
-        ps = con.prepareStatement(s1);
-        ps.setString(1, params.get("gid"));
-        ps.executeUpdate();
-        return "Success!";
+        if (!rs.next()) {
+            throw new Exception("There isnt any Tag with ID: " + params.get("{gid}"));
+        } else {
+            String s2 = "delete from tag_checklist where Tg_id = ?";
+            ps = con.prepareStatement(s2);
+            ps.setString(1, params.get("{gid}"));
+            ps.executeUpdate();
+
+            s2 = "delete from tag where Tg_id = ?";
+            ps = con.prepareStatement(s2);
+            ps.setString(1, params.get("{gid}"));
+            ps.executeUpdate();
+            return "Success!";
+        }
     }
 
     @Override
